ai_service/classifier.py
import random
from PIL import Image

class CivicClassifier:
    def __init__(self):
        # Load your model here (e.g., TensorFlow, PyTorch)
        # self.model = load_model("pothole_detector.h5")
        print("Model initialized (Mock Mode)")
        self.categories = ["Pothole", "Garbage", "Broken Streetlight", "Normal"]

    def predict(self, image_path):
        """
        Takes an image path/stream and predicts the issue category.
        """
        try:
            # Preprocess image
            img = Image.open(image_path)
            img = img.resize((224, 224))
            
            # Run inference (Mocking logic based on random for now)
            # In real scenario: prediction = self.model.predict(img)
            
            # Simple simulation: 
            # If we had a model, we would return the class with highest probability
            
            return random.choice(self.categories)
        except Exception as e:
            print(f"Error during prediction: {e}")
            return "Unknown"
