ai_service/requirements.txt
flask
flask-cors
pillow
pymongo
dnspython
python-dotenv
gunicorn
